move artifact::decl to its own file
